:fire: move: 스케줄러 초안 이동
- 리팩토링 전후를 비교하자 패키지 새로 생성 및 이동
