+++ src/main/java/com/project/cheerha/domain/notice/dto/draft/JobOpeningKeywordDto.java
@@ -1,12 +1,13 @@
-package com.project.cheerha.domain.notice.dto;
+package com.project.cheerha.domain.notice.dto.draft;
 
 import com.querydsl.core.annotations.QueryProjection;
 
 /**
  * 스케줄러에 채용 공고 관련 데이터를 가져오는 데 필요한 DTO
+ *
  * @param jobOpeningId 채용 공고 ID
- * @param keywordId 키워드 ID
- * @param url 채용 공고의 URL
+ * @param keywordId    키워드 ID
+ * @param url          채용 공고의 URL
  */
 public record JobOpeningKeywordDto(
     Long jobOpeningId,
